// src/models/website/WebLeads.model.ts
export type WebLeadsListResponseType = {
    _id: string
    order_id: string | null
    name: string
    phone: string
    email: string
    address: string
    address1: string
    landmark: string
    city: string
    state: string
    country: string
    zip_code: string
    quantity: string
    remark: string
    sdate: string
    idtag: string
    product_name: string
    mode: string
    paymeny_mode: string
    url: string
    price: string
    leadStatus: string
    isDeleted: boolean
    isActive: boolean
    createdAt: Date
    updatedAt: Date
    __v: number
    companyCode: string
    leadType: string
    paymentGatewayName: string
    paymentMode: string
    paymentStatus: string
    transactionId: string
    companyId: string
}